Decompile Library Background Parallax Entities
Decompiles background bookcase, chairs & desk, & shadows behind the
ectoplasms.

Split library background into a separate segment.
