+++ config/symbols.us.stlib.txt
@@ -57,9 +57,12 @@ bone_projectile_velocity_x = 0x80182CF4;
 sensors_special = 0x80182D24;
 sensors_move = 0x80182D2C;
 g_UnkPrimHelperRot = 0x801ADDA4;
-ST0_EntityBackgroundBlock = 0x801ADED0;
+LIB_EntityBackgroundBlock = 0x801ADED0;
 PlayerIsWithinHitbox = 0x801ADF98;
 func_801B0AA4 = 0x801AE01C;
+EntityBackgroundBookcase = 0x801AF280;
+EntityBackgroundFurniture = 0x801AF538;
+EntityBackgroundShadow = 0x801AF7B8;
 EntityLibrarianChair = 0x801B037C;
 CutsceneUnk1 = 0x801B95EC;
 SetCutsceneScript = 0x801B9640;
